requirements.txt
# 🦊 Fennec Customer Support — Dependencies

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ── Fennec RAG Framework ──────────────────────────────────────────────────────
fennec-rag>=1.0.0

# ── LLM & Embeddings ─────────────────────────────────────────────────────────
openai>=1.40.0
anthropic>=0.34.0              # Optional: for Claude support
google-generativeai>=0.8.0     # Optional: for Gemini support

# ── Vector Database ───────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Sentence Transformers (Semantic Router) ───────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.0.0

# ── Document Loaders ──────────────────────────────────────────────────────────
pymupdf>=1.24.0        # PDF loading
python-docx>=1.1.0     # DOCX loading
beautifulsoup4>=4.12.0 # HTML loading
lxml>=5.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.8.0
httpx>=0.27.0
aiofiles>=23.0.0
